[Common Code] Clean up some of the code

+++ Source/Common/File Class.cpp
@@ -91,8 +91,7 @@ bool CFile::Open(const char * lpszFileName, uint32_t nOpenFlags)
         dwCreateFlag = OPEN_EXISTING;
 
     // attempt file creation
-    HANDLE hFile = ::CreateFile(lpszFileName, dwAccess, dwShareMode, &sa,
-        dwCreateFlag, FILE_ATTRIBUTE_NORMAL, NULL);
+    HANDLE hFile = ::CreateFile(lpszFileName, dwAccess, dwShareMode, &sa, dwCreateFlag, FILE_ATTRIBUTE_NORMAL, NULL);
     if (hFile == INVALID_HANDLE_VALUE)
     { //#define ERROR_PATH_NOT_FOUND             3L
         //ULONG err = GetLastError();
@@ -169,12 +168,12 @@ uint32_t CFile::Read(void* lpBuf, uint32_t nCount)
         return 0;   // avoid Win32 "null-read"
     }
 
-    ULONG dwRead = 0;
+    DWORD dwRead = 0;
     if (!::ReadFile(m_hFile, lpBuf, nCount, &dwRead, NULL))
     {
         return 0;
     }
-    return (UINT)dwRead;
+    return (uint32_t)dwRead;
 }
 
 long CFile::Seek(long lOff, SeekPosition nFrom)
